# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Promote Api",
        "description": "Handsmoe API",
        "version": "1.0.0"
    },
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/artists/{artist}": {
            "get": {
                "summary": "Get an artist.",
                "parameters": [
                    {
                        "name": "artist",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "x-example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get an artist",
                        "schema": {
                            "items": {
                                "$ref": "#/definitions/Artist"
                            }
                        }
                    },
                    "404": {
                        "description": "Artist not found"
                    }
                }
            },
            "put": {
                "summary": "Update an Artist.",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "description": "Artist to update",
                        "schema": {
                            "$ref": "#/definitions/ArtistType"
                        }
                    },
                    {
                        "name": "artist",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "x-example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artist updated",
                        "schema": {
                            "$ref": "#/definitions/Artist"
                        }
                    },
                    "400": {
                        "description": "Invalid Request"
                    },
                    "404": {
                        "description": "Artist not found"
                    }
                }
            },
            "patch": {
                "summary": "Pacth an Artist.",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "description": "Artist to patch",
                        "schema": {
                            "$ref": "#/definitions/ArtistType"
                        }
                    },
                    {
                        "name": "artist",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "x-example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artist updated",
                        "schema": {
                            "$ref": "#/definitions/Artist"
                        }
                    },
                    "400": {
                        "description": "Invalid Request"
                    },
                    "404": {
                        "description": "Artist not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete an Artist.",
                "parameters": [
                    {
                        "name": "artist",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "x-example": "2"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Artist deleted"
                    },
                    "404": {
                        "description": "Artist not found"
                    }
                }
            }
        },
...
        "/gigs/{gig}": {
            "get": {
                "summary": "Get a gig.",
                "parameters": [
                    {
                        "name": "gig",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "x-example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get a gig",
                        "schema": {
                            "$ref": "#/definitions/Gig"
                        }
                    },
                    "404": {
                        "description": "Gig not found"
                    }
                }
            },
            "put": {
                "summary": "Modify an existing Gig.",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "description": "Gig to add",
                        "schema": {
                            "$ref": "#/definitions/GigType"
                        }
                    },
                    {
                        "name": "gig",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "x-example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gig updated",
                        "schema": {
                            "$ref": "#/definitions/Gig"
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "Gig not found"
                    }
                }
            },
            "patch": {
                "summary": "Patch an existing Gig.",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "description": "Gig to add",
                        "schema": {
                            "$ref": "#/definitions/GigType"
                        }
                    },
                    {
                        "name": "gig",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "x-example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gig patched",
                        "schema": {
                            "$ref": "#/definitions/Gig"
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "Gig not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete an existing Gig.",
                "description": "ApiDoc(statusCodes = {\n  204 = \"Gig deleted\",\n  404 = \"Gig not found\"\n})",
                "parameters": [
                    {
                        "name": "gig",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "x-example": "2"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Gig deleted"
                    },
                    "404": {
                        "description": "Gig not found"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Artist": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "image_name": {
                    "type": "string"
                }
            },
            "type": "object"
        },
        "ArtistType": {
            "required": [
                "name",
                "slug",
                "bio",
                "gigs",
                "labels"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "gigs": {
                    "example": "[1, 2, 3]",
                    "type": "array",
                    "format": "[AppBundle\\Entity\\Gig id]"
                },
                "labels": {
                    "example": "[1, 2, 3]",
                    "type": "array",
                    "format": "[AppBundle\\Entity\\Label id]"
                }
            },
            "type": "object"
        },
...
        "Gig": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string",
                    "format": "date-time"
                },
                "end_date": {
                    "type": "string",
                    "format": "date-time"
                },
                "venue": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "facebook_link": {
                    "type": "string"
                }
            },
            "type": "object"
        },
        "GigType": {
            "required": [
                "name",
                "startDate",
                "endDate",
                "venue",
                "address",
                "facebookLink"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string",
                    "format": "date-time"
                },
                "endDate": {
                    "type": "string",
                    "format": "date-time"
                },
                "venue": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "facebookLink": {
                    "type": "string"
                }
            },
            "type": "object"
        }
    }
}